requirements.txt
langchain_google_genai==2.1.8
langgraph==0.5.4
pydantic==2.11.7
python-dotenv==1.1.1
typing_extensions==4.14.1
